+++ src/main/java/com/pulbatte/pulbatte/global/controller/TokenController.java
@@ -3,7 +3,7 @@ package com.pulbatte.pulbatte.global.controller;
 import com.pulbatte.pulbatte.global.dto.MsgResponseDto;
 import com.pulbatte.pulbatte.global.dto.RequestToken;
 import com.pulbatte.pulbatte.global.exception.SuccessCode;
-import com.pulbatte.pulbatte.global.jwt.TokenDto;
+import com.pulbatte.pulbatte.global.dto.TokenDto;
 import com.pulbatte.pulbatte.global.security.UserDetailsImpl;
 import com.pulbatte.pulbatte.global.service.TokenService;
 import lombok.RequiredArgsConstructor;
