+++ src/main/java/com/pulbatte/pulbatte/global/service/TokenService.java
@@ -1,13 +1,11 @@
 package com.pulbatte.pulbatte.global.service;
 
-import com.pulbatte.pulbatte.global.dto.MsgResponseDto;
 import com.pulbatte.pulbatte.global.dto.RequestToken;
 import com.pulbatte.pulbatte.global.entity.RefreshToken;
 import com.pulbatte.pulbatte.global.exception.CustomException;
 import com.pulbatte.pulbatte.global.exception.ErrorCode;
-import com.pulbatte.pulbatte.global.exception.SuccessCode;
 import com.pulbatte.pulbatte.global.jwt.JwtUtil;
-import com.pulbatte.pulbatte.global.jwt.TokenDto;
+import com.pulbatte.pulbatte.global.dto.TokenDto;
 import com.pulbatte.pulbatte.global.repository.RefreshTokenRepository;
 import com.pulbatte.pulbatte.user.entity.User;
 import com.pulbatte.pulbatte.user.repository.UserRepository;
