+++ src/main/java/com/pulbatte/pulbatte/user/controller/KakaoUserController.java
@@ -1,8 +1,7 @@
 package com.pulbatte.pulbatte.user.controller;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
-import com.pulbatte.pulbatte.global.jwt.TokenDto;
-import com.pulbatte.pulbatte.post.dto.KakaoUserInfoDto;
+import com.pulbatte.pulbatte.global.dto.TokenDto;
 import com.pulbatte.pulbatte.user.service.KakaoService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
